requirements.txt
absl-py==2.3.0
accelerate==1.8.1
adversarial-robustness-toolbox==1.20.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.4
annotated-types==0.7.0
attrs==25.3.0
charset-normalizer==3.4.2
colorlog==6.9.0
dendropy==5.0.8
einops==0.8.1
fair-esm==2.0.0
filelock==3.13.1
frozenlist==1.7.0
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.12
gitpython==3.1.45
greenlet==3.2.3
grpcio==1.73.0
hf-xet==1.1.5
huggingface-hub==0.34.3
idna==3.10
jinja2==3.1.4
joblib==1.5.1
lightning==2.5.2
lightning-utilities==0.14.3
mako==1.3.10
markdown==3.8.2
markupsafe==2.1.5
mpmath==1.3.0
multidict==6.5.0
networkx==3.3
nltk==3.9.1
numpy==1.26.1
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
optuna==4.4.0
packaging==25.0
pandas-stubs==2.2.3.250527
pillow==11.0.0
propcache==0.3.2
protobuf==6.31.1
pydantic==2.11.7
pydantic-core==2.33.2
pydeprecate==0.3.1
pytorch-lightning==2.5.2
rotary-embedding-torch==0.8.6
scikit-learn==1.5.0
scipy==1.15.3
sentence-transformers==2.6.1
sentencepiece==0.2.0
sentry-sdk==2.34.1
smmap==5.0.2
sqlalchemy==2.0.41
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchmetrics==1.7.3
torchvision==0.22.1+cu118
transformers==4.54.1
triton==3.3.1
typing-inspection==0.4.1
urllib3==2.5.0
wandb==0.21.0
werkzeug==3.1.3
yarl==1.20.1